Split into common defs/utils
