+++ common/lib/common-definitions/src/logger.ts
@@ -138,96 +138,5 @@ export interface IFluidError {
     addDetails: (props: ITelemetryProperties, debugData: ISensitiveDebugData) => void;
 }
 
-export const isIFluidError = (err: any): err is IFluidError => (
-    typeof(err?.errorType) === "string" &&
-    typeof(err?.getFluidTelemetryProps) === "function" && //* Or only check this for brevity/perf?
-    typeof(err?.getSensitiveDebugData) === "function" &&
-    typeof(err?.addDetails) === "function");
-
-// ////////////////////// fluidError.ts in common-utils /////////////////////////////
-
-export abstract class FluidError extends Error implements IFluidError {
-    public abstract errorType: string;
-
-    private props: ITelemetryProperties = {};
-    private debugData: ISensitiveDebugData = {};
-    constructor(
-        message: string,
-        props: ITelemetryProperties = {},
-        debugData: ISensitiveDebugData = {},
-    ) {
-        super(message);
-        this.addDetails(props, debugData);
-    }
-
-    public getFluidTelemetryProps() { return { ...this.props, message: this.message, errorType: this.errorType }; }
-    public getSensitiveDebugData() {
-        return { ...this.debugData, stack: this.stack ?? "" };
-    } //* Or implement deep copy? Not even possible...?
-
-    public addDetails(props: ITelemetryProperties, debugData: Partial<ISensitiveDebugData>) {
-        this.props = { ...this.props, ...props };
-        this.debugData = { ...this.debugData, ...debugData };
-    }
-}
-
-export class GenericError extends FluidError {
-    public errorType: string = "generic";
-}
-
-export function wrapAsFluidError(err: any): IFluidError {
-    if (isIFluidError(err)) {
-        return err;
-    }
-
-    //* start with promoting err's message, per present behavior, and then pull out in later scoped change.
-    //* Same with stack above?
-    return new GenericError(
-        "External Error",
-        {},
-        { innerError: err },
-    );
-}
-
-// ///////////////// Testing ////////////////////
-
-export enum DriverErrorType {
-    throttlingError = "throttlingError",
-    asdf = "asdf",
-}
-
-export interface IDriverErrorBase extends IFluidError {
-    readonly errorType: DriverErrorType;
-    canRetry: boolean;
-    online?: string;
-}
-
-export interface IThrottlingWarning extends IDriverErrorBase {
-    readonly errorType: DriverErrorType.throttlingError;
-    readonly retryAfterSeconds: number;
-}
-
-export class ThrottlingError extends FluidError implements IThrottlingWarning {
-    readonly errorType = DriverErrorType.throttlingError;
-    readonly canRetry = true;
-
-    constructor(
-        errorMessage: string,
-        readonly retryAfterSeconds: number,
-        statusCode?: number,
-    ) {
-        //* As-is, need to include retryAfterSeconds here since we're no longer
-        //* just pulling every property off the object but rather keeping them
-        //* separated in props. Something to consider.
-        super(errorMessage, { retryAfterSeconds, statusCode });
-    }
-}
-
-export class AssertError extends FluidError {
-    constructor(
-        errorMessage: string,
-        readonly errorType: string,
-    ) {
-        super(errorMessage);
-    }
-}
+export const isIFluidError = (err: any): err is IFluidError =>
+    typeof(err?.getFluidTelemetryProps) === "function";
